+++ src/ConnectionManager.h
@@ -26,7 +26,10 @@ class ConnectionManager final {
   }
 
   std::shared_ptr<ConnectionHolder> connect(ip_addr_t *ipaddr, uint16_t port);
-  bool listen(uint16_t port);
+
+  // Listens on a port. The `reuse` parameter controls the SO_REUSEADDR flag.
+  // This returns whether the attempt was successful.
+  bool listen(uint16_t port, bool reuse);
 
   bool isListening(uint16_t port);