+++ src/QNEthernetUDP.h
@@ -26,8 +26,12 @@ class EthernetUDP final : public UDP {
   ~EthernetUDP();
 
   // Start listening on a port. This returns true if successful and false if the
-  // port is in use.
+  // port is in use. This calls begin(localPort, false).
   uint8_t begin(uint16_t localPort) override;
+
+  // Start listening on a port and set the SO_REUSEADDR socket option according
+  // to the `reuse` parameter. This returns whether the attempt was successful.
+  uint8_t begin(uint16_t localPort, bool reuse);
 
   uint8_t beginMulticast(IPAddress ip, uint16_t port) override;
   void stop() override;